requirements.txt
websockets==8.1
pyinstaller==4.0
pycryptodome==3.9.8
pywin32==228
Pillow==7.2.0
pyscreenshot==2.2
pyimgur==0.6.0